change transactions and transaction types PKs

+++ db-scripts/2019-9-4_12-38-create-tansactions-table.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS transaction_types 
 (
-    id int PRIMARY KEY AUTO_INCREMENT,
+    id tinyint unsigned PRIMARY KEY AUTO_INCREMENT,
     name varchar(50) NOT NULL,
     date_created TIMESTAMP NOT NULL default CURRENT_TIMESTAMP,
     date_last_edited TIMESTAMP on update CURRENT_TIMESTAMP
@@ -8,13 +8,13 @@ CREATE TABLE IF NOT EXISTS transaction_types
 
 CREATE TABLE IF NOT EXISTS transactions 
 (
-    id int PRIMARY KEY AUTO_INCREMENT,
-    user_id int NOT NULL,
-    account_id int NOT NULL,
-    type int NOT NULL,
+    id int unsigned PRIMARY KEY AUTO_INCREMENT,
+    user_id int unsigned NOT NULL,
+    account_id int unsigned NOT NULL,
+    type tinyint unsigned NOT NULL,
     amount double NOT NULL,
     notes varchar(100) default NULL,
-    journal_entry int NOT NULL,
+    journal_entry int unsigned NOT NULL,
     corrected boolean default false,
     server_generated boolean default false,
     date_created TIMESTAMP NOT NULL default CURRENT_TIMESTAMP,
